+++ cmd/cove/main.go
@@ -1967,11 +1967,11 @@ func handleNetworkCommand(args []string) {
 	case "help":
 		fmt.Println(NetworkModeHelp())
 	case "audit":
-		if len(args) != 2 {
-			fmt.Fprintln(os.Stderr, "Usage: cove network audit <run-id>")
-			os.Exit(1)
+		if len(args) > 1 && isHelpArg(args[1]) {
+			printNetworkAuditUsage(os.Stdout)
+			return
 		}
-		if err := PrintNetworkAudit(os.Stdout, args[1]); err != nil {
+		if err := RunNetworkAudit(os.Stdout, args[1:]); err != nil {
 			fmt.Fprintf(os.Stderr, "error: %v\n", err)
 			os.Exit(1)
 		}